WNB-181: Move Menu and Loading components to typescript
