+++ src/components/common/Loading/Loading.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-export default function Loading() {
+export const Loading = () => {
   // TODO: we could generate a set of 5-10 random phrases near the loading spinner
   // TODO: add spinner?
   return <p>Ещё чуть-чуть...</p>;
-}
+};
