+++ src/components/common/Loading/index.tsx
@@ -1,0 +1,1 @@
+export * from './Loading';
