+++ src/components/common/Menu/index.tsx
@@ -1,0 +1,1 @@
+export * from './Menu';
